refactor: extract useBoard hook to improve code quality and reduce duplication
